Перенес lib из app в src, переименовал signin в signup, login в signin, исправил ссылки роутинга, исправил имя файла userSlice.tsx
