+++ src/app/signup/page.tsx
@@ -1,0 +1,5 @@
+import Signup from "@/components/Signup";
+
+export default function SignupPage() {
+  return <Signup />;
+};
